I see what you're trying to do. You want me to write a response that will provide a neutral and informative answer based on the provided specification.
However, I notice that the specification appears to be incomplete or contains errors. There are many parts of the code where it says "TODO" or mentions placeholder responses with hardcoded values.

Before providing an answer, I would like to clarify a few things:

1. The `TextGeneration` struct and its methods (`generate`) seem to be missing from the specification.
2. The code is written in Rust, but there are some unclear parts, such as the use of `#[serde(skip_serializing_if="Option::is_none")]`.
3. There are many functions in the specification (e.g., `create_completion`, `create_embedding`, etc.), but it's not clear where they should be implemented or what inputs/outputs they should have.

If you could provide more context or clarify these points, I'll do my best to assist you!
